requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
asyncio>=3.4.3
openai>=0.27.0
psutil>=5.9.0
aiohttp>=3.8.0
schedule>=1.2.0
docker>=6.1.0
PyYAML>=6.0
numpy>=1.24.0
seaborn>=0.12.0
matplotlib>=3.6.0
python-dotenv>=1.0.0
textblob>=0.19.0